+++ apps/api/src/models/companies/graphql/companies.resolver.ts
@@ -22,7 +22,9 @@ export class CompaniesResolver {
     @Args('createCompanyInput') args: CreateCompanyInput,
     @GetUser() user: GetUserType,
   ) {
-    // checkRowLevelPermission(user, args.uid);
+    const managerId = args.managerId;
+
+    checkRowLevelPermission(user, managerId);
     return this.companiesService.create(args);
   }
 
@@ -44,8 +46,12 @@ export class CompaniesResolver {
   ) {
     const company = await this.prisma.company.findUnique({
       where: { id: args.id },
+      include: { Managers: true },
     });
-    // checkRowLevelPermission(user, company.uid);
+    checkRowLevelPermission(
+      user,
+      company.Managers.map((man) => man.uid),
+    );
     return this.companiesService.update(args);
   }
 
@@ -55,8 +61,14 @@ export class CompaniesResolver {
     @Args() args: FindUniqueCompanyArgs,
     @GetUser() user: GetUserType,
   ) {
-    const company = await this.prisma.company.findUnique(args);
-    // checkRowLevelPermission(user, company.uid);
+    const company = await this.prisma.company.findUnique({
+      ...args,
+      include: { Managers: true },
+    });
+    checkRowLevelPermission(
+      user,
+      company.Managers.map((man) => man.uid),
+    );
     return this.companiesService.remove(args);
   }
 }
