+++ apps/api/src/models/reviews/graphql/reviews.resolver.ts
@@ -22,7 +22,7 @@ export class ReviewsResolver {
     @Args('createReviewInput') args: CreateReviewInput,
     @GetUser() user: GetUserType,
   ) {
-    // checkRowLevelPermission(user, args.uid)
+    checkRowLevelPermission(user, args.customerId);
     return this.reviewsService.create(args);
   }
 
@@ -45,7 +45,7 @@ export class ReviewsResolver {
     const review = await this.prisma.review.findUnique({
       where: { id: args.id },
     });
-    // checkRowLevelPermission(user, review.uid)
+    checkRowLevelPermission(user, review.customerId);
     return this.reviewsService.update(args);
   }
 
@@ -56,7 +56,7 @@ export class ReviewsResolver {
     @GetUser() user: GetUserType,
   ) {
     const review = await this.prisma.review.findUnique(args);
-    // checkRowLevelPermission(user, review.uid)
+    checkRowLevelPermission(user, review.customerId);
     return this.reviewsService.remove(args);
   }
 }
