+++ apps/api/src/models/slots/graphql/slots.resolver.ts
@@ -1,6 +1,6 @@
 import { Resolver, Query, Mutation, Args } from '@nestjs/graphql';
 import { SlotsService } from './slots.service';
-import { Slot } from './entity/slot.entity';
+import { ReturnCount, Slot } from './entity/slot.entity';
 import { FindManySlotArgs, FindUniqueSlotArgs } from './dtos/find.args';
 import { CreateSlotInput } from './dtos/create-slot.input';
 import { UpdateSlotInput } from './dtos/update-slot.input';
@@ -18,12 +18,54 @@ export class SlotsResolver {
 
   @AllowAuthenticated()
   @Mutation(() => Slot)
-  createSlot(
+  async createSlot(
     @Args('createSlotInput') args: CreateSlotInput,
     @GetUser() user: GetUserType,
   ) {
-    // checkRowLevelPermission(user, args.uid)
+    const garage = await this.prisma.garage.findUnique({
+      where: { id: args.garageId },
+      include: { Company: { include: { Managers: true } } },
+    });
+    checkRowLevelPermission(
+      user,
+      garage.Company.Managers.map((manager) => manager.uid),
+    );
     return this.slotsService.create(args);
+  }
+  @AllowAuthenticated('manager')
+  @Mutation(() => ReturnCount)
+  async creatManySlots(
+    @Args('createSlotInput') args: CreateSlotInput,
+    @Args('count', {
+      type: () => Number,
+    })
+    count: number,
+    @GetUser() user: GetUserType,
+  ) {
+    const garage = await this.prisma.garage.findUnique({
+      where: { id: args.garageId },
+      include: {
+        Company: {
+          include: { Managers: true },
+        },
+      },
+    });
+    checkRowLevelPermission(
+      user,
+      garage.Company.Managers.map((manager) => manager.uid),
+    );
+    const typeCount = await this.prisma.slot.count({
+      where: {
+        garageId: args.garageId,
+        type: args.type,
+      },
+    });
+
+    const slots = Array.from({ length: count }).map((num, index) => ({
+      ...args,
+      displayName: `${args.type} ${typeCount + index + 1}`,
+    }));
+    return this.prisma.slot.createMany({ data: slots });
   }
 
   @Query(() => [Slot], { name: 'slots' })
@@ -42,8 +84,22 @@ export class SlotsResolver {
     @Args('updateSlotInput') args: UpdateSlotInput,
     @GetUser() user: GetUserType,
   ) {
-    const slot = await this.prisma.slot.findUnique({ where: { id: args.id } });
-    // checkRowLevelPermission(user, slot.uid)
+    const slot = await this.prisma.slot.findUnique({
+      where: { id: args.id },
+      include: {
+        Garage: {
+          include: {
+            Company: {
+              include: { Managers: true },
+            },
+          },
+        },
+      },
+    });
+    checkRowLevelPermission(
+      user,
+      slot.Garage.Company.Managers.map((man) => man.uid),
+    );
     return this.slotsService.update(args);
   }
 
@@ -53,8 +109,22 @@ export class SlotsResolver {
     @Args() args: FindUniqueSlotArgs,
     @GetUser() user: GetUserType,
   ) {
-    const slot = await this.prisma.slot.findUnique(args);
-    // checkRowLevelPermission(user, slot.uid)
+    const slot = await this.prisma.slot.findUnique({
+      where: { id: args.where.id },
+      include: {
+        Garage: {
+          include: {
+            Company: {
+              include: { Managers: true },
+            },
+          },
+        },
+      },
+    });
+    checkRowLevelPermission(
+      user,
+      slot.Garage.Company.Managers.map((man) => man.uid),
+    );
     return this.slotsService.remove(args);
   }
 }
